CK_travel_and_work/src/test/java/cz/cvut/fel/rsp/travelandwork/service/CategoryServiceTest.java
package cz.cvut.fel.rsp.travelandwork.service;

import cz.cvut.fel.rsp.travelandwork.environment.util.Generator;
import cz.cvut.fel.rsp.travelandwork.exception.NotFoundException;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityManager;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

@RunWith(SpringRunner.class)
@SpringBootTest
@Transactional
@TestPropertySource(locations = "classpath:application-test.properties")
public class CategoryServiceTest {
    Category testCategory = Generator.generateCategory();
    List<Category> list = Arrays.asList(Generator.generateCategory(), Generator.generateCategory(), Generator.generateCategory());

    @Autowired
    TranslateService ts;

    @Autowired
    private EntityManager em;

    @Autowired
    private CategoryService sut;

    @Before
    public void init() {
        sut.create(testCategory);
        for (Category cs: list) {
            sut.create(cs);
            System.out.println(cs.toString());
        }
    }

    @Test
    @Transactional
    @Rollback
    public void findAll_findsAllCategories() {
        assertNotNull(sut.findAll());
    }

    @Test
    @Transactional
    @Rollback
    public void find_findCategoryById() throws NotFoundException {
        assertEquals(testCategory, sut.find(testCategory.getId()));
    }

    @Test
    @Transactional
    @Rollback
    public void create_createsNewCategory() throws NotFoundException {
        Category newCategory = Generator.generateCategory();
        sut.create(newCategory);
        assertEquals(newCategory, sut.find(newCategory.getId()));
    }

    @Test
    @Transactional
    @Rollback
    public void update_updatesCategory() throws NotFoundException {
        testCategory.setName("No Named Achievement. haha.");
        sut.update(testCategory.getId(), testCategory);
        assertEquals(testCategory, sut.find(testCategory.getId()));
    }


}
